requirements.txt
# Core dependencies for RAG system
langchain==0.1.20
langchain-openai==0.1.7
langchain-core==0.1.53
openai==1.30.1
qdrant-client==1.9.1
sentence-transformers==5.0.0

# Streamlit UI dependencies
streamlit==1.47.1

# Environment and configuration
python-dotenv==1.0.1

# Optional dependencies for reranking
torch==2.7.1
transformers==4.54.0

# Development and testing
pytest==8.4.1

# Data handling
pandas==2.0.3
numpy==1.24.4